added if statement to toggle username/login display
