+++ client/src/actions/listActions.js
@@ -39,6 +39,7 @@ export function getsingleCat(slug,id) {
   })
 }
 
+
 export function getSearch(searchResults) {
   axios.get('/search/' + searchResults).then(resp => {
     store.dispatch({
