+++ client/src/components/App.js
@@ -13,11 +13,12 @@ import Scat from './Singlecat'
 import Form from './Form'
 import About from './About'
 import Search from './Search'
-import SignIn from './SignIn'
+import LogIn from './LogIn'
 import Register from './Register'
 import ComingSoon from './ComingSoon'
 import ContactUs from './ContactUs'
-import { withAuth } from '../lib/auth';
+
+// import { withAuth } from '../lib/auth';
 import Chat from './Chat'
 
 class App extends Component {
@@ -28,17 +29,17 @@ class App extends Component {
           <div>
             <Search />
             <Switch>
-              <Route path="/signin" component={SignIn} />
+              <Route path="/login" component={LogIn} />
               <Route path="/register" component={Register} />
               <Route exact path="/"  component={Home} />
               <Route path="/about" component={About} />
-              <Route path="/post/:id"  component={Post} />
               <Route path="/chatroom" component={Chat} />
               <Route path="/posts/:slug/:id"  component={Scat} />
               <Route path="/contactus"  component={ContactUs} />
-              <Route path="/form/:slug/:id" component={Form} />
               <Route path="/:slug/:id"  component={List} /> 
               <Route path="/comingSoon" component={ComingSoon} />
+              <Route path="/post/:id"  component={Post} />
+              <Route path="/form/:slug/:id" component={Form} />
             </Switch>
           </div>
         </Router>
@@ -49,4 +50,3 @@ class App extends Component {
 
 export default App
 
-//<Route  exact path="/"  component={Signin} />
