+++ client/src/components/Footer.js
@@ -1,8 +1,6 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import { Link } from 'react-router-dom'
-import ScrollUpButton from 'react-scroll-up-button'
-
 import '../styles/footerStyles.css'
