+++ client/src/components/List.js
@@ -6,6 +6,8 @@ import { Link } from 'react-router-dom'
 import '../styles/listStyles.css'
 
 import Footer from './Footer'
+
+
 class List extends Component {
   componentDidMount() {
     getPosts(this.props.match.params.slug, this.props.match.params.id)
@@ -16,7 +18,6 @@ class List extends Component {
 
       <div className="listCon">
         <div className="bodyCon">
-          <h1>Crop Trade App</h1>
           <h1 className="catname">{this.props.match.params.slug}</h1> 
           <div className="bodyCon">
               
@@ -28,7 +29,7 @@ class List extends Component {
               <div className="gridCon">
                   {this.props.posts.map(item => (
                     <div key={item.id} className="grid">
-                       <Link  className="postLi" to={`/post/${item.id}`}> <img className="gridimage" src={item.photo}/><p>{item.title} (Las Vegas)</p> </Link> 
+                       <Link  className="postLi" to={`/post/${item.id}`}> <img alt="noPic" className="gridimage" src={item.photo}/><p>{item.title} (Las Vegas)</p> </Link> 
                     </div>
                   ))}
               </div>
