+++ client/src/components/Register.js
@@ -32,7 +32,7 @@ class Register extends Component {
             return(
                 <div>
                 <div id="login">
-                    <h2>Choose a username to be recognized!</h2>
+                    <h2>Choose a Username and Password</h2>
                     <form onSubmit={this.handleSubmit} id="loginForm">
                         <input name='username' className="loginStuff" type="text" value={this.state.username} onChange={this.handleChange} placeholder="Enter a Username"/>
                         <input name='password' className="loginStuff" type="text" value={this.state.password} onChange={this.handleChange} placeholder="Enter a Password"/>
