+++ client/src/styles/loginStyles.css
@@ -20,3 +20,11 @@
   transition: all 0.3s ease 0s;
   outline: none;
   }
+
+  #login {
+    margin-bottom: 25%
+  }
+
+  #registerLink {
+    margin-left: 50px;
+  }
