+++ server/routes/auth.js
@@ -38,7 +38,10 @@ Router.post('/login', (req, res, next)=>{
 })
 
 
+// WORKS WITH POSTMAN, DONT MESS WITH
+
 Router.post('/register', (req, res, next) => {
+    // console logs response to the terminal, see username and pw from inputs come through in res
     console.log("authRoute post", res)
     const sql = 'INSERT INTO users (username, password) VALUES (?,?) '
     if (!req.body.username || !req.body.password){
